Plot Cumulative Intervals

This function generates a cumulative interval plot based on the provided lower and upper bounds, cluster IDs, and other optional parameters.

Usage

plot_intervals_cumulative(
  lower,
  upper,
  cluster_id,
  truth = NA,
  min,
  max,
  facet_wrap = NULL,
  weighted = NULL,
  show_quantiles = TRUE,
  ncol = 3
)

Arguments

lower

A numeric vector of lower bounds.

upper

A numeric vector of upper bounds.

cluster_id

A vector of cluster IDs corresponding to the intervals.

truth

A numeric vector of ground truth values. Default is NA.

min

The minimum value for the x-axis.

max

The maximum value for the x-axis.

facet_wrap

A logical value indicating whether to use facet wrapping. Default is FALSE.

weighted

An optional vector of weights for the intervals.

show_quantiles

A logical value indicating whether to show quantiles on the plot. Default is TRUE.

ncol

The number of columns for facet wrapping. Default is 2.

Value

A ggplot object representing the cumulative interval plot.