Verbal Quantifier Data

A subset of data from the data collected by Kloft & Heck (2024) containing the probability interval ratings for verbal quantifiers. The dataset is in the long format where all responses for the lower interval bound and upper interval bound, respectively, are in one column.

Usage

quantifiers

Format

quantifiers

A data frame with 3,344 rows and 8 columns:

id_person

Unique identifier for each person

id_item

Unique identifier for each item

name_ger

German name of the quantifier

name_eng

English name of the quantifier

truth

True value of the quantifier if applicable

scale_min

Minimum value of the scale

scale_max

Maximum value of the scale

width_min

Minimum possible width of the interval rating

x_L

Lower bound of the interval rating

x_U

Upper bound of the interval rating

Source

https://osf.io/67vyj/

References

Kloft, M., & Heck, D. W. (2024). Discriminant validity of interval response formats: Investigating the dimensional structure of interval widths. Educational and Psychological Measurement, 0 (0). https://doi.org/10.1177/00131644241283400